API developing on MarcaController.php.

+++ app/Http/Controllers/MarcaController.php
@@ -25,49 +25,82 @@ class MarcaController extends Controller {
      * Store a newly created resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function store(Request $request) {
-        $marca = $this->marca->create($request->all());
-        return $marca;
+        $request->validate($this->marca->rules(), $this->marca->feedback());
+
+        $imagem = $request->file('imagem');
+        $imagem_urn = $imagem->store('imagens', 'public');
+
+        $marca = $this->marca->create([
+            'nome' => $request->nome,
+            'imagem' => $imagem_urn
+        ]);
+
+        return response()->json($marca, 201);
     }
 
     /**
      * Display the specified resource.
      *
      * @param Integer $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function show($id) {
         $marca = $this->marca->find($id);
-        return $marca;
+        if ($marca === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
+        return response()->json($marca);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Marca $marca
-     * @return \Illuminate\Http\Response
+     * @param Integer $id
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function update(Request $request, Marca $marca) {
-        /*
-        print_r($request->all()); //os dados atualizados
-        echo '<hr>';
-        print_r($marca->getAttributes()); //os dados antigos
-        */
+    public function update(Request $request, $id) {
+        $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        if ($request->method() === 'PATCH') {
+            $regrasDinamicas = [];
+            //percorrendo todas as regras definidas no Model
+            foreach ($marca->rules() as $input => $regra) {
+                //coletar apenas as regras aplicáveis aos parâmetros parciais da requisição PATCH
+                if (array_key_exists($input, $request->all())) {
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+            $request->validate($regrasDinamicas, $marca->feedback());
+        } else {
+            $request->validate($marca->rules(), $marca->feedback());
+        }
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param \App\Models\Marca $marca
-     * @return \Illuminate\Http\Response
+     * @param Integer $id
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function destroy(Marca $marca) {
+    public function destroy($id) {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi removida com sucesso!'];
+        return response()->json(['msg' => 'A marca foi removida com sucesso!']);
     }
 }
